// src/app/models/team-member.ts
// export interface TeamMember {
//   location: string;
//   name: string;
//   position: string;
//   skills: string[];
//   profilePicture: string;
//   project: string;

  
//   // Add any other properties relevant to a team member
// }


export interface TeamMember {
  name: string;
  position: string;
  skills: string[];
  profilePicture: string;
  location: string;
  projects: string[];
  email: string; 
}